Began Text Clarity Correction, Custom Buy Amount
Added a button and box to let players choose a custom amount to buy.

Created a system which attempts to correct text colors if they are too similar to the background by changing their HSV Value. Compares those values for all text and tries to make sure every color is at least 40 Value away from it, unless the color cannot become any lighter or darker, in which case it simply fails to accomplish anything.

Not implemented in all places. Only on the home button costs, tooltip, recycler, and grotto dipose buttons, and gate, because the others are a little more complicated.

Will go back and clean up the code to make it less awful to look at. There are bugs at the moment, I am aware.