New Income System, Rebalance
Redid a piece of the income system and optimized the RK function. The system keep track of both the base income and the upgrade multipliers to it, and keeps track of the final number after applying them, allowing computations to skip having to access and account for those multipliers when calculating incomes.

Changed a few values on a few upgrades. Tried to increase fun factor of abandoned by adding a very powerful filter upgrade at the end. Also made agriculture cost less clicks on the manual sponge button.

Added new upgrade, "Superior Search Algorithms" to make better the start of ancient parts progression.