Misc
Lots of things in this push: Added tooltip for resources in the resource table (needs work), removed gaps between cells in resource table to accomodate the new tooltip, added simplified mode to grotto, began adding column headers to the grotto. Also started work on new haven and added a table of all real, actual incomes after costscaling is applied, and the inverse of that table as well.