Nearly Done with Modifier Rework
Completed the modifier rework for upgrades and world modifiers, as far as i know, with the possible exclusion of some world modifiers. Artifacts are not integrated yet.

Grotto table has a simplified mode now. It's going to need some work, but there's a system in place to dynamically create column headers regardless of the size of the table. Removed the separate resource boost column from the grotto, changed upgrades color to the old boost modifier color (from weird gray to nicer green).

Actions is gonna yell at me for not linting this. I'm not particularly concerned, I'll lint it later!