Changed Upgrade System
Added a check which pretty much makes the invisible property obsolete to fix a bug and a lot of annoying side-effects. The way it works: if the list of upgrades for a world doesnt contain an upgrade's name, then it doesn't even bother adding it. As a side effect which I do not mind, at minimum one line must be present for every upgrade in a world. FYI, still significantly smaller than before.

I don't actually know if it makes the invisible property obsolete or not, I'll decide whether or not to remove it tomorrow.