Cool Title Bar
This commit replaces the gigantic title picture + buttons with a short title bar + buttons, which makes the game way nicer looking.

This needs an option to switch between the old and new modes, and I've made sure to (mostly) leave room for that implementation by accomplishing it with only new CSS (and one small change to the body margin).