Aspect Sprites, Aspect 'Unlock Conditions', Essence Counter
Added sprites for a bunch of aspects. The renderer now automatically tries to find a sprite with a matching label in sprites.js.

Added a new method to each aspect which allows for new, special unlock conditions. The tree colors locked nodes and lines going to them differently.

Added an essence counter to the aspect tree screen.

Tried to separate these into different commits, but they got all tangled up, so I couldn't.