Changed how Aspects are Applied
Aspects are now applied through a function in SharkGame.ArtifactTree, taken from SharkGame.Gateway.

The aspect apply method now takes an argument: a time, similar to a handlingTime, which specifies if its effects are being applied on levelup or on game/world initialization. Differentiating between these cases is important for various applications.