Implemented Relative Luminance Checking
Phew, this was a hell of a challenge. The game can now estimate relative luminance using a custom quadratic estimation formula and solve for the proper adjustment to meet contrast requirements.

Even then, we're not meeting minimum contrast requirements. We need to change all the colors of the backgrounds and whatnot to be darker, methinks.