Added New Aspects
Functionalities not implemented yet. This is nearly all of them, just a few more to go!

Locked some aspects. E.g. Patience is not useful because you have to beat 3 more worlds for it to be useful, and once you beat 3 worlds and get a payout, you're already done with the demo. Destiny gamble also isn't useful because there needs to be at least 4 worlds for rerolling to be justified in any way.