Haven Tweaks
Put a bandaid on the problems with haven's game design. Will address them at a later time, there are more pressing issues at the moment.

Updated lots of flavor text and prices in haven to better match the theme and story, and smooth over the pacing.

Fixed choruses still generating essence. That was not intentional.