Default Sidebar Width Changed
Increased default sidebar width to 30%. This should make the game more compatible with smaller screen sizes.

I'm pretty sure that my screen is so much bigger because I have two monitors, one higher res than the other - so the UI scales to the size of the smaller screen, leaving extra room on the bigger screen.