Cheaty Powers
This is a prototype of sorts for a visual implementation of a sort of debug menu on its own tab. This also moves CheatsAndDebug to its own file.

Hopefully, this helps to clean up index.html.

Still need to add a way to make the tab disappear. Use cad.debug() to enable the tab & debug button.