Destiny Gamble, Flags
Implemented the Flags system. Flags are just a special name for global variables that are stored on save and restored on load.

There are two kinds: flags and persistent flags. Regular flags are cleared every time the player enters the gateway (aka beats or loses a world). Persistent flags are only cleared on save wipe.

The example given here is the destiny gamble aspect, which stores how many rolls it has using persistent flags.

The save updater needs an update, it will get one when I finish updating the updater in the dev branch.