Cool Sprites!!!!!!!! Tooltip Changes! Other stuff!!!!
Made the sprites for the marker way cooler.

Changed tooltips to not have fade-in time when dealing with the marker.

Added proper formatting to minute hand, formatted its text/log entries, gave it tons of flavor messages for turning on/off.

Added tooltips for mousing over the marker slots. Just generally did a lot.