Grotto Change
Total resources are no longer a measurement in the grotto for whether or not a resource should be there. Now, it's whether or not it's being produced or consumed by anything. This has the effect of making things which produce and consume tar appear in the grotto even if you haven't ever actually had tar go above 0.