Fixed Bug with Empty Cost Text
Changed the formula for cost functions to all make use of the difference between what you own and what you want. This is due to the discovery of a mathematical simplification of the linear cost formula which uses the difference between the current and desired amounts to bypass precision limits, and is also a shorter (and very likely faster) formula.