Correct log height and auto scroll.
- The bottom of the scroll bar on chrome at least is not there because
the log's parent div has a max height. However, setting the log to have
a max height fixes it. Also means that the user is unable to scroll to
the last message.
- I noticed when playing that if I scrolled down on the log, it never
went back to the top. With this change, it scrolls up 1 extra item
for each item pushed to it. It also clips so the entirety of each
message is visible when it automatically scrolls upward.